+++ libft/ft_array_from_lst.c
@@ -6,7 +6,7 @@
 /*   By: sherbert <sherbert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/10 11:53:27 by sherbert          #+#    #+#             */
-/*   Updated: 2021/10/10 12:18:32 by sherbert         ###   ########.fr       */
+/*   Updated: 2021/10/10 14:04:38 by sherbert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void    ft_array_from_lst(int **a, t_list *lst)
     i = 0;
     while (lst->next)
 	{
-		a[i] = (int)lst->content;
+		a[i] = lst->content;
         i++;
         lst = lst->next;
 	}
